+++ src/utils/generateStages.cpp
@@ -39,8 +39,9 @@ matjson::Value generateStages(std::vector<float> sps)
           .from = from1,
           .to = to1,
           .checked = false,
-          .note = "",
           .attempts = 0,
+          .note = "",
+          .completedAt = 0,
           .completionCounter = 0});
     }
 
@@ -58,8 +59,9 @@ matjson::Value generateStages(std::vector<float> sps)
             .from = from2,
             .to = to2,
             .checked = false,
-            .note = "",
             .attempts = 0,
+            .note = "",
+            .completedAt = 0,
             .completionCounter = 0});
       }
     }